dashboard/requirements.txt
# Enhanced Dashboard Requirements (builds on existing backend)
# Install these additional packages for the dashboard features

# Web framework for enhanced dashboard
flask==2.3.3
flask-cors==4.0.0

# Machine learning for stock prediction
scikit-learn==1.3.0
statsmodels==0.14.0

# News API integration
newsapi-python==0.2.7

# Alpha Vantage for stock search
alpha-vantage==2.3.1

# Enhanced caching
requests-cache==1.1.1

# Additional visualization
bokeh==3.2.1

# For Render deployment compatibility
gunicorn==21.2.0

# For enhanced Streamlit integration
plotly==5.15.0

# Additional data processing
numpy==1.24.3
pandas==2.0.3
